JSONPluginMgr: add backend-specific options and support for a preferred backend
This adds two new API constructs.

The first is the notion of the 'preferred' JSON backend.
The preferred backend is set by calling, for example,
jsonpickle.json.set_preferred_backend('simplejson')
This ensures that the the simplejson backend is tried before any
other backend.

User-supplied backends can be added with the load_backend() method.
jsonpickle comes with support for cjson, json, simplejson and
demjson by default.

The second construct is allowing backend-specific encoder options.
This is done by calling, for example,
jsonpickle.json.set_encoder_options('simplejson', indent=4)

The result of this is that jsonpickle calls simplejson.dumps
with the indent=4 option.

Signed-off-by: David Aguilar <davvid@gmail.com>

git-svn-id: https://jsonpickle.googlecode.com/svn/trunk@17 ba5bf794-d048-0410-8f42-e1fdefefffeb