jsonpickle: add support for tuples and sets
jsonpickle now supports tuples and sets.  Previously all three
collection types (list, tuple, set) mapped to the same JSON
data type and thus we lost the type information when coming
back from JSON.

We now track tuples and sets separately so that we can construct
an object of the appropriate type when unpickling.

Signed-off-by: David Aguilar <davvid@gmail.com>

git-svn-id: https://jsonpickle.googlecode.com/svn/trunk@39 ba5bf794-d048-0410-8f42-e1fdefefffeb