encode: add a max_depth parameter
max_depth allows one to limit the depth of recursion in jsonpickle.
This is a slightly modified version of the patch provided by
Antonin Hildebrand.  The modification was to allow max_depth=0.
Also, the original used patch None to denote an unlimited depth.
We now use negative numbers to denote this instead of None.

The _isunpickleable function in pickler.py was removed since
it was easier to just move the 'unpicklable' option into the
keyword arguments of encode().

Signed-off-by: David Aguilar <davvid@gmail.com>

git-svn-id: https://jsonpickle.googlecode.com/svn/trunk@43 ba5bf794-d048-0410-8f42-e1fdefefffeb