Add support for backends that are stored in sub-modules
This ensures that we can load a backend regardless of where
it lives in the module namespace.  We did not originally
support loading a backend from, e.g. foo.bar.baz -- only
backends that live in a top-level module were allowed.

We now support loading backends from any module, be it a
top-level module or a sub-module.

A test case was added to ensure that this behavior is
preserved.

Signed-off-by: David Aguilar <davvid@gmail.com>

git-svn-id: https://jsonpickle.googlecode.com/svn/trunk@46 ba5bf794-d048-0410-8f42-e1fdefefffeb