pluginmgr: make load_backend() exception-safe
load_backend() could get things in an incomplete state if
any of its arguments triggered exceptions.

We now guard against all exceptions and rollback the
internal state if any steps in load_backend fail.

Test cases were added to ensure that the internal validity
of the plugin manager are maintained.

Signed-off-by: David Aguilar <davvid@gmail.com>

git-svn-id: https://jsonpickle.googlecode.com/svn/trunk@52 ba5bf794-d048-0410-8f42-e1fdefefffeb