Better support for cyclical data structures
One limitation of the initial "referenced object ID" implementation was
that it relied on a special "xquery"-like "/foo/bar/baz" object name
notation.  This was a simple mapping that could only map into objects
and their attributes; it fell apart when the object it needed to
reference first appeared as an element of a list.

jsonpickle now uses an ID-based object reference notation.
The pickler keeps track of objects it has seen and assigns each
a unique number as it is traversed.  That object's JSON representation
is represented by, e.g. {"py/id": 42} later when that same object is
seen elsewhere in an object graph.

This method is robust enough to index into list elements because the
lookup is id-based instead of an ad-hoc query format.

The pickler no longer writes the old {"py/ref": "/foo/bar"} reference
dicts, it only writes {"py/id": int} references.

The unpickler, though, understands both notations and can still unpickle
JSON created by earlier versions of jsonpickle.  This keeps backwards
compatibility so that we can still read files written by earlier versions
of the pickler.

Closes #12

Signed-off-by: David Aguilar <davvid@gmail.com>