tests: simplify tests by moving samples.py to jsonpickle._samples
jsonpickle requires that the objects it reconstitutes are available
globally via a module.

This makes some of the doctests difficult to write because samples.py
is not always available on the sys.path.

Simplify things by moving our samples objects to jsonpickle/_samples.py.
This removes some duplication from the doctests and ensures that
nosetests --with-doctest works correctly without needing to modify
the PYTHONPATH.

Signed-off-by: David Aguilar <davvid@gmail.com>