pickler: handle object references with int dict keys
Integer dict keys causes the lexical sorting to be different
from the integer sorting, which leads to IndexErrors when the
dictionary contains references to objects.

Avoid the problem by using strings as the canonical
representation when sorting.  This ensures that a consistent
sort order is preserved internally and by the JSON backends.

Closes #54

Reported=by: Jason R. Coombs <jaraco@jaraco.com>
Signed-off-by: David Aguilar <davvid@gmail.com>