pickler: simplify the fix for #55
9fd2884e342bc7094b2839140a4b6713daa1e877 was a little too complex
because the unpickler was dealing with the fallout of _mkref() getting
called twice in the flatten() code path for objects with custom
handlers.

The simpler solution is to not call _mkref() twice.  We only need to
call it on the original object instance; calling it on the flattened
representation was the source of the original bug.

This is a much better solution.

Closes #55

Signed-off-by: David Aguilar <davvid@gmail.com>