backend: allow disabling of the fallthrough-on-error behavior
By default, jsonpickle tries the next backend when decoding or
encoding using a backend fails.

This can make it difficult to force jsonpickle to use a specific
backend, and catch errors, because the error will be suppressed
and may not be raised by the subsequent backend.

Add an enable_fallthrough() option that allows us to disable
this behavior.

Signed-off-by: David Aguilar <davvid@gmail.com>