pickler: add support for defaultdicts with self as their default factory
A defaultdict subclass can set `self` as its default_factory.
This causes an infinite recursion error within jsonpickle
because we do not notice the cycle between the object instance
and its default factory.

Teach the pickler to recognize when a default_factory has been
seen before so that it can emit a reference.  This makes these
objects picklable.

Teach the unpickler to defer trying to load the factory until
after the proxy object is available for use by referencing.
This makes it possible to unpickle these objects.

Add a `__call__()` implementation to the _Proxy class so that we
can guarantee that reconstructed objects' default_factory is
callable and constructs the right class.

Closes #106
Reported-by: @rako233 via github.com
Helped-by: Marcin Tustin <marcin.tustin@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>