unpickler: prevent proxies from escaping via defaultdict.default_factory
Check whether defaultdict's default_factory contains a proxy,
and if so, replace it with the final instance.  Add tests to
ensure that proxies cannot escape via the default_factory.

The original fix for #106 modified _Proxy to support __call__()
but that change is no longer needed since the proxy will no
longer escape into that context.

Related-to: #106
Signed-off-by: David Aguilar <davvid@gmail.com>