tox.ini: execute tests from the tests/ directory
59754692c5c5271fb1e9aa835030b32c8337b89f removed the chdir
directive, but it is useful because it allows us to catch
packaging errors by forcing the jsonpickle that is run
by the tests to be the jsonpickle that was installed
via tox (rather than the one from the source tree).

Bring back the chdir directive to prevent Python from
finding the jsonpickle in the current directory.

We still specify `jsonpickle` and `.` to the nosetests
invocation so that the jsonpickle doctests are executed.

Helped-by: Ivan Smirnov <i.s.smirnov@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>