pickler: better bytes handling with stdlib json
The stdlib python2.7 json module does not want to see bytes objects in
its object stream.

Check for bytes earlier so that they are not handled by the
"is_primitive()" code path.  This ensures that bytes are decoded
into unicode strings on Python2 when the bytes stream is utf-8,
and into a base64-representation otherwise.

Correctly detect when data needs to be encoded and decoded in
b64{encode,decode}.

Closes #217
Helped-by: Jason R. Coombs <jaraco@jaraco.com>
Signed-off-by: David Aguilar <davvid@gmail.com>