unpickler: better support for sqlalchemy >= 1.3
The _safe_hasattr() workaround for older versions of SQLAlchemy
surprisingly breaks newer versions.

Be optimistic and catch the runtime error so that newer versions do not
incur the cost of the original workaround.  The older versions are still
supported since they'll land in the old code path.

This is justifiable since it's a generic code path and will speed up
things for 99% of the other objects out there, including newer
sqlalchemy versions.

Closes #254
Signed-off-by: David Aguilar <davvid@gmail.com>