jsonpickle: improve serialization for non-string keys
Object ID references can get mixed up when an object is used
as both the key and a value.  This happens because the iteration
order is different after serializing to json due to the json://
prefix that is added to these keys.

This can be solved by special-casing these entries so that we
can process them separately.  Flatten and restore is now a
two-phase encode/decode approach.  The first phase encodes the
simple key/value pairs only, and the second phase encodes the
entries with object keys only.

This ensures that object ID references are generated and decoded
in a consistent order.

Closes #286
Signed-off-by: David Aguilar <davvid@gmail.com>