jsonpickle: preserve dict order on Python3
Retain dict order now that dict iteration is guaranteed to be
deterministic and order-preserving in Python3.

Set sort_keys=True on Python2 only, and use a util wrapper
function to paper over the differences in iteration style.

While we're here, improve Python2's performance by avoiding a
copy through the use of iteritems() -- we have a wrapper
function so we might as well specialize it now that it's
convenient to do so.

Closes #193
Signed-off-by: David Aguilar <davvid@gmail.com>