compat: support python3.4 + 3.5
Older Python3 versions did not guarantee dict key order.
Adjust the code to reflect that we can only really expect
this behavior in Python3.6+.

Although the 3.6 release notes state that this behavior
should not be relied upon, Python3.7 makes this an official
language feature, so this is totally crescent fresh.

https://docs.python.org/3/whatsnew/3.6.html#new-dict-implementation