enums: properly serialize multiple enums when make_refs=False
When the same enum is encountered under the make_refs=False code path we
(wrongly) assumed that the object was introducing a cycle, and thus could
potentially lead to infinite recursion during traversal.

Treat enums similarly to atoms in that they are no longer considered
when detecting cycles.  This allows the make_refs=False code path to
correctly serialize enums when the same enum value is encountered
multiple times during object traversal.

Resolves #235
Reported-by: @pmmcmullen94 via github.com
Reported-by: Thomas Baier <tbaier@math.bme.hu>
Signed-off-by: David Aguilar <davvid@gmail.com>