pickler: create copies of seen objects
Previously the make_refs=False / unpicklable=False code path
would avoid emitting data for objects it had already seen.
It mostly did this to avoid cycles, as its reference
system is disabled in these modes. The end result was that null
was emitted for any objects that appear more than once in the
object graph.

Improve the behavior by keeping track of objects that we
have already flattened and return the pre-flattened
representation when we re-encounter the same object.

This is a simple solution that fixes encoding of datetime
objects with UTC tzinfo in these scenarios.  It also improves
encoding of objects with multiple references to the same
object as the object will now be properly encoded each
time it is encountered in the object graph.

Resolves #333 #334 #337
Signed-off-by: David Aguilar <davvid@gmail.com>