pickler: flatten Cython functions using _flatten_function()
Newer versions of Cython create functions that provide a __reduce__()
implementation that does not play nicely with the strategies used
by jsonpickle.

Detect Cython functions and record references to them directly
so that they are properly pickled. This fixes the ability to
encode and decode pandas objects.

We avoid recording "py/ref" references to Cython functions
because they throw off the ref offsets and are not needed.
"py/function" references, when unpickled, resolve to module
globals that exist outside of the object graph.

Fixes: #460