pickler/unpickler: pickle functions and Cython functions earlier
Cython v3 produces functions that pass the util.is_object() tests.
This complicated the unpickler because it needed to start tracking
refs for functions, which it never needed to do before.

Expand the meaning of is_module_function() so that it includes cython
functions and use this check earlier in the pickling process.
This makes the pickler handle Cython functions before they ever make
it into the "if util.is_object(obj):" code path.

These changes restore the pickler's original semantics by ensuring
that functions are excluded from the referencing machinery completely.

Related-to: #460