jsonpickle: add support for top-level tuples with cyclical references
Tuples were not previously part of the referencing machinery. This made
it so that references to tuples would not get restored -- copies were
getting created instead.

Leverage the "mutuple" package to make it so that jsonpickle can edit
tuples after they are created. This allows us to replace proxies in
tuples similarly to how lists are updated when proxies are replaced.