Fix data types in tests for NumPy 2
Fix a number of test compatibility issues with NumPy 2.  That is:

- replace `np.float_` with `np.float64`,
- replace `np.complex_` with `np.complex128`,
- remove `np.unicode_` as equivalent to `np.str_`.

The LHS types are also aliases to the RHS types in NumPy 1.x, so this
is NFC on NumPy 1.x.  These aliases were removed in 2.x, so the code
needs to be updated for tests to pass on NumPy 2.x.