tags: use py/mod tags for modules instead of py/repr
Make things clearer and 1 byte smaller by using a "py/mod" tag
when pickling modules.

The older py/repr is still supported by the reader, but we do
not emit it.

Now that the repr tags and modules are separated then we can
make the older eval() behavior opt-in instead of opt-out.
We do this by setting the "safe" option to True by default,
thus requiring users that want the backwards-compatible
set "safe=False" when unpickling.