ci: remove github action that creates commits (#518)
Remove the write permission and the committing of code so that github
bots are no longer creating commits. That was an interesting setup, but
it also leads to surprises where github's branch will advance beyond
what we push.

Run black before isort since errors in black have a higher chance
of occuring.