backend: load ujson before yaml
84b928e08f54ef6d0ea8d7fd91cb98cffec0ab13 added the new `yaml` backend
before the existing `ujson` backend, which was an oversight as it can
change behavior for users that are expecting the `ujson` backend to be
used when both are available.

Make jsonpickle behave more closely to its original behavior by loading
`yaml` after `ujson`. The new `yaml` backend will only be used after all
of the existing backends are attempted.

Related-to: #528