backend: make the yaml backend opt-in
YAML is a much more lenient format so it can cause jsonpickle to load
files that are expected to fail when interpreted using a JSON reader.
Let users control whether or not the yaml backend is used.

Move registration of the yaml backend into a separate
jsonpickle.ext.yaml module. Do not register the yaml module by default
so that users can opt-in to it.

Closes: #550
Reported-by: Jason R. Coombs @jaraco