unpickler: avoid holding on to external references
Even though the unpickler goes out of scope immediately after decode()
completes, the Python garbage collector is not able to determine that
garbage collection can occur despite the fact that no references are
being held. The observed behavior is that Python is not deleting
the ephemeral unpickler instance.

Make this easier for the GC to handle by explicitly resetting the
ephemeral Unpickler instance at the end of decode(). This guarantees
that no external references are being held and allows them to
be cleaned up in a slightly more predictable fashion.

The test case was added to the numpy test suite because triggering
this behavior currently requires the numpy extension to be used.

Closes: #579